* fix endless recursion in to hither unused function mulle_objc_universeid_is_sane * fix an assert in class-search, that curiously was never triggered added some more mostly superflous tests, but the more the merrier. Reduced scope means its easier to find compiler bugs

+++ test/version/check_version.c
@@ -39,23 +39,23 @@ int   main( int argc, const char * argv[])
               MULLE_OBJC_RUNTIME_VERSION_PATCH);   
    if( MULLE_OBJC_RUNTIME_VERSION != version)
    {
-      fprintf( stderr, "Runtime header runtime version/compiler version mismatch (%u %u)",
-         version, MULLE_OBJC_RUNTIME_VERSION);
+      fprintf( stderr, "Runtime header runtime version/compiler version mismatch (%u %lu)",
+         version, (unsigned long) MULLE_OBJC_RUNTIME_VERSION);
       return( 1);
    }
 
    version = ((JIT_MAJOR << 20) | (JIT_MINOR << 8) |JIT_PATCH);
    if( MULLE_OBJC_RUNTIME_VERSION != version)
    {
-      fprintf( stderr, "JIT/Runtime header runtime version mismatch (%u %u)\n",
-         version, MULLE_OBJC_RUNTIME_VERSION);
+      fprintf( stderr, "JIT/Runtime header runtime version mismatch (%u %lu)\n",
+         version, (unsigned long) MULLE_OBJC_RUNTIME_VERSION);
       return( 1);
    }
 
    if( MULLE_OBJC_RUNTIME_LOAD_VERSION != JIT_LOAD_VERSION)
    {
-      fprintf( stderr, "JIT/Runtime header load version mismatch (%u %u)\n", 
-         JIT_LOAD_VERSION, MULLE_OBJC_RUNTIME_LOAD_VERSION);
+      fprintf( stderr, "JIT/Runtime header load version mismatch (%u %lu)\n",
+         JIT_LOAD_VERSION, (unsigned long) MULLE_OBJC_RUNTIME_LOAD_VERSION);
       return( 1);
    }
    return 0;
